fix the two remaining scintilla x64 warnings
git-svn-id: https://notepad2-mod.googlecode.com/svn/trunk@564 28bd50df-7adb-d945-0439-6e466c6a13cc

+++ scintilla/lexers/LexPerl.cxx
@@ -248,7 +248,7 @@ static bool styleCheckSubPrototype(LexAccessor &styler, unsigned int bk) {
 
 static bool isMatch(const char *sref, char *s) {
 	// match per-line delimiter - must kill trailing CR if CRLF
-	int i = strlen(s);
+	int i = static_cast<int>(strlen(s));
 	if (i != 0 && s[i - 1] == '\r')
 		s[i - 1] = '\0';
 	return (strcmp(sref, s) == 0);
